Fix timeout data type bugs. (#131)
This commit resolves a couple problems.

First, it changes the return type of NoteGetMs. NoteGetMs is simply a wrapper
around the user-supplied millisecond timer hook. The prototype for this hook
specifies that it should return a uint32_t yet NoteGetMs's return type is long
unsigned int. This commit changes the return type of NoteGetMs to uint32_t.

Second, there are many places in the code where we are comparing and assigning
millisecond timer values (i.e. gotten via NoteGetMs) to size_t variables. This
is a bug. On some platforms (e.g. the Arduino Uno), size_t is 16 bits, so
assigning one of these 32-bit millisecond timer values to a 16-bit variable can
lead to undefined behavior when the timer value is larger than what can fit into
16 bits. This commit fixes all these bugs by using uint32_t instead of size_t.

+++ n_request.c
@@ -303,7 +303,7 @@ J *NoteRequestResponseWithRetry(J *req, uint32_t timeoutSeconds)
 */
 char * NoteRequestResponseJSON(const char *reqJSON)
 {
-    size_t transactionTimeoutMs = (CARD_INTER_TRANSACTION_TIMEOUT_SEC * 1000);
+    uint32_t transactionTimeoutMs = (CARD_INTER_TRANSACTION_TIMEOUT_SEC * 1000);
     char *rspJSON = NULL;
 
     if (reqJSON == NULL) {
@@ -476,7 +476,7 @@ J *noteTransactionShouldLock(J *req, bool lockNotecard)
     //   - If the request is a `web.*`, follow the same logic, but instead
     //     of using the standard timeout, use the Notecard timeout of 90
     //     seconds for all `web.*` transactions.
-    size_t transactionTimeoutMs = (CARD_INTER_TRANSACTION_TIMEOUT_SEC * 1000);
+    uint32_t transactionTimeoutMs = (CARD_INTER_TRANSACTION_TIMEOUT_SEC * 1000);
 
     // Interrogate the request
     if (JContainsString(req, (reqType ? "req" : "cmd"), "note.add")) {
